feat: add read-only store
Closes #13
