+++ junk_drawer/__init__.py
@@ -1,5 +1,6 @@
 """Top-level module for junk_drawer."""
 from .store import Store
+from .read_store import ReadStore
 
-__all__ = ["Store"]
+__all__ = ["Store", "ReadStore"]
 __version__ = "0.1.0"
