+++ pyproject.toml
@@ -15,8 +15,8 @@ flake8-docstrings = "^1.5.0"
 mock = "^4.0.2"
 mypy = "^0.782"
 pytest = "^6.1.0"
-pytest-watch = "^4.2.0"
 pytest-asyncio = "^0.14.0"
+pytest-xdist = "^2.2.0"
 flake8-annotations = "^2.4.1"
 
 [tool.black]
