+++ tests/test_read_store.py
@@ -3,7 +3,7 @@ import pytest
 from pathlib import PurePath
 from mock import AsyncMock  # type: ignore[attr-defined]
 
-from junk_drawer import Store
+from junk_drawer import ReadStore
 from junk_drawer.errors import ItemDecodeError, ItemAccessError
 from junk_drawer.filesystem import (
     DirectoryEntry as DirEntry,
@@ -19,7 +19,7 @@ pytestmark = pytest.mark.asyncio
 
 
 async def test_store_exists_with_nonexistent_key(
-    store: Store[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
 ) -> None:
     """It should return False from store.exists if file doesn't exist."""
     mock_filesystem.file_exists.return_value = False
@@ -30,7 +30,7 @@ async def test_store_exists_with_nonexistent_key(
 
 
 def test_store_exists_sync_with_nonexistent_key(
-    store: Store[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
 ) -> None:
     """It should return False from store.exists if file doesn't exist."""
     mock_filesystem.sync.file_exists.return_value = False
@@ -41,7 +41,7 @@ def test_store_exists_sync_with_nonexistent_key(
 
 
 async def test_store_exists_with_existing_key(
-    store: Store[CoolModel], mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], mock_filesystem: AsyncMock
 ) -> None:
     """It should return True from store.exists if file exists."""
     mock_filesystem.file_exists.return_value = True
@@ -51,7 +51,7 @@ async def test_store_exists_with_existing_key(
 
 
 def test_store_exists_sync_with_existing_key(
-    store: Store[CoolModel], mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], mock_filesystem: AsyncMock
 ) -> None:
     """It should return True from store.exists if file exists."""
     mock_filesystem.sync.file_exists.return_value = True
@@ -61,7 +61,7 @@ def test_store_exists_sync_with_existing_key(
 
 
 async def test_store_get_with_nonexistent_key(
-    store: Store[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
 ) -> None:
     """It should return None from store.get if file doesn't exist."""
     mock_filesystem.read_json.side_effect = PathNotFoundError()
@@ -74,7 +74,7 @@ async def test_store_get_with_nonexistent_key(
 
 
 def test_store_get_sync_with_nonexistent_key(
-    store: Store[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
 ) -> None:
     """It should return None from store.get if file doesn't exist."""
     mock_filesystem.sync.read_json.side_effect = PathNotFoundError()
@@ -87,7 +87,7 @@ def test_store_get_sync_with_nonexistent_key(
 
 
 async def test_store_get_returns_read_result(
-    store: Store[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
 ) -> None:
     """It should return a model from store.get if file exists."""
     result = CoolModel(foo="bar", bar=42)
@@ -101,7 +101,7 @@ async def test_store_get_returns_read_result(
 
 
 def test_store_get_sync_returns_read_result(
-    store: Store[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
 ) -> None:
     """It should return a model from store.get if file exists."""
     result = CoolModel(foo="bar", bar=42)
@@ -114,8 +114,36 @@ def test_store_get_sync_returns_read_result(
     )
 
 
+async def test_store_get_returns_read_result_from_deep_key(
+    store: ReadStore[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
+) -> None:
+    """It should return a model from store.get if deep file exists."""
+    result = CoolModel(foo="bar", bar=42)
+    mock_filesystem.read_json.return_value = result
+    item = await store.get("foo/bar")
+
+    assert item == result
+    mock_filesystem.read_json.assert_called_with(
+        store_path / "foo" / "bar", parse_json=store.parse_json
+    )
+
+
+def test_store_get_sync_returns_read_result_from_deep_key(
+    store: ReadStore[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
+) -> None:
+    """It should return a model from store.get if deep file exists."""
+    result = CoolModel(foo="bar", bar=42)
+    mock_filesystem.sync.read_json.return_value = result
+    item = store.get_sync("foo/bar")
+
+    assert item == result
+    mock_filesystem.sync.read_json.assert_called_with(
+        store_path / "foo" / "bar", parse_json=store.parse_json
+    )
+
+
 async def test_store_get_all_keys_reads_dir(
-    store: Store[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
 ) -> None:
     """It should read the directory and return all JSON files as keys."""
     mock_filesystem.read_dir.return_value = ["foo", "bar", "baz"]
@@ -126,7 +154,7 @@ async def test_store_get_all_keys_reads_dir(
 
 
 def test_store_get_all_keys_sync_reads_dir(
-    store: Store[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
 ) -> None:
     """It should read the directory and return all JSON files as keys."""
     mock_filesystem.sync.read_dir.return_value = ["foo", "bar", "baz"]
@@ -137,7 +165,7 @@ def test_store_get_all_keys_sync_reads_dir(
 
 
 async def test_store_get_all_entries_reads_dir(
-    store: Store[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
 ) -> None:
     """It should reads for all files in the directory to get all entries."""
     mock_filesystem.read_json_dir.return_value = [
@@ -156,7 +184,7 @@ async def test_store_get_all_entries_reads_dir(
 
 
 def test_store_get_all_entries_sync_reads_dir(
-    store: Store[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
 ) -> None:
     """It should reads for all files in the directory to get all entries."""
     mock_filesystem.sync.read_json_dir.return_value = [
@@ -175,7 +203,7 @@ def test_store_get_all_entries_sync_reads_dir(
 
 
 async def test_store_get_all_items_reads_dir(
-    store: Store[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
 ) -> None:
     """It should read all files in the directory to get all items."""
     mock_filesystem.read_json_dir.return_value = [
@@ -194,7 +222,7 @@ async def test_store_get_all_items_reads_dir(
 
 
 async def test_store_get_all_items_sync_reads_dir(
-    store: Store[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], store_path: PurePath, mock_filesystem: AsyncMock
 ) -> None:
     """It should read all files in the directory to get all items."""
     mock_filesystem.sync.read_json_dir.return_value = [
@@ -213,7 +241,7 @@ async def test_store_get_all_items_sync_reads_dir(
 
 
 async def test_store_get_raises_invalid_json(
-    store: Store[CoolModel], mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], mock_filesystem: AsyncMock
 ) -> None:
     """It should raise a validation error if the data fails to parse."""
     mock_filesystem.read_json.side_effect = FileParseError("oh no")
@@ -235,7 +263,7 @@ async def test_store_get_raises_invalid_json(
 
 
 async def test_store_get_raises_read_error(
-    store: Store[CoolModel], mock_filesystem: AsyncMock
+    store: ReadStore[CoolModel], mock_filesystem: AsyncMock
 ) -> None:
     """It should raise an access error if the data fails to read."""
     mock_filesystem.read_json.side_effect = FileReadError("oh no")
@@ -257,7 +285,7 @@ async def test_store_get_raises_read_error(
 
 
 async def test_silenced_validation_error(
-    ignore_errors_store: Store[CoolModel],
+    ignore_errors_store: ReadStore[CoolModel],
     store_path: PurePath,
     mock_filesystem: AsyncMock,
 ) -> None:
